[tests] Testing for empty ATG init
Immplements tests for initializing an empty ATG and also tests that
indexing an empty ATG fails as expected. It doesn't test the xmin or
xmax since those are liable to change (maybe we want None or negative
values to tip off it's empty) and handling times should probably be the
user's responsibility.